Move MacM1 settings to its own file
